requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
click==8.2.1
colorama==0.4.6
coverage==7.10.2
Django==5.2.5
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg==3.2.9
psycopg-binary==3.2.9
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
ruff==0.12.8
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
